+++ jcloud/helper.py
@@ -8,16 +8,19 @@ import tempfile
 import threading
 import warnings
 from contextlib import contextmanager
-from distutils.version import LooseVersion
 from pathlib import Path
-from typing import Union
+from typing import Dict, Union
 from urllib.parse import urlparse
 from urllib.request import Request, urlopen
 
 import pkg_resources
 import yaml
+from packaging.version import Version
 from rich import print
+from rich.highlighter import ReprHighlighter
 from rich.panel import Panel
+
+from .env_helper import EnvironmentVariables, expand_dict
 
 __windows__ = sys.platform == 'win32'
 
@@ -30,7 +33,7 @@ def _version_check(package: str = None, github_repo: str = None):
         if not github_repo:
             github_repo = package
 
-        cur_ver = LooseVersion(pkg_resources.get_distribution(package).version)
+        cur_ver = Version(pkg_resources.get_distribution(package).version)
         req = Request(
             f'https://pypi.python.org/pypi/{package}/json',
             headers={'User-Agent': 'Mozilla/5.0'},
@@ -41,7 +44,7 @@ def _version_check(package: str = None, github_repo: str = None):
             j = json.load(resp)
             releases = j.get('releases', {})
             latest_release_ver = list(
-                sorted(LooseVersion(v) for v in releases.keys() if '.dev' not in v)
+                sorted(Version(v) for v in releases.keys() if '.dev' not in v)
             )[-1]
             if cur_ver < latest_release_ver:
                 print(
@@ -67,17 +70,18 @@ def is_latest_version(package: str = None, github_repo: str = None) -> None:
     threading.Thread(target=_version_check, args=(package, github_repo)).start()
 
 
-def get_logger():
+def get_logger(name='jcloud'):
     from rich.logging import RichHandler
 
-    logging.basicConfig(
-        level=os.environ.get('JCLOUD_LOGLEVEL', 'DEBUG'),
-        format='%(message)s',
-        datefmt='[%X]',
-        handlers=[RichHandler(rich_tracebacks=True)],
-    )
+    logger = logging.getLogger(name)
+    logger.setLevel(os.environ.get('JCLOUD_LOGLEVEL', 'INFO'))
 
-    return logging.getLogger('jcloud')
+    rich_handler = RichHandler(rich_tracebacks=True)
+    formatter = logging.Formatter('%(message)s')
+    rich_handler.setFormatter(formatter)
+    logger.addHandler(rich_handler)
+
+    return logger
 
 
 def get_or_reuse_loop():
@@ -168,12 +172,14 @@ def valid_uri(uses):
         return False
 
 
-def normalized(path: Union[str, Path]):
+def normalized(path: Union[str, Path], envs: Dict):
     _normalized = True
     with open(path) as f:
         _flow_dict = yaml.safe_load(f.read())
 
     if 'executors' in _flow_dict:
+        with EnvironmentVariables(envs):
+            expand_dict(_flow_dict, context=envs)
         for executor in _flow_dict['executors']:
             uses = executor.get('uses', None)
             if uses is None:
@@ -183,3 +189,13 @@ def normalized(path: Union[str, Path]):
             else:
                 _normalized = False
     return _normalized
+
+
+class CustomHighlighter(ReprHighlighter):
+    highlights = ReprHighlighter.highlights + [
+        r"(?P<url>(grpc|grpcs)://[-0-9a-zA-Z$_+!`(),.?/;:&=%#]*)"
+    ]
+
+
+def remove_prefix(s: str, prefix: str):
+    return s[len(prefix) :] if s.startswith(prefix) else s