+++ jcloud/parsers/base.py
@@ -2,6 +2,8 @@ import argparse
 
 
 def set_base_parser():
+    import os
+
     from .. import __version__
     from .helper import _chf, colored
 
@@ -14,13 +16,13 @@ def set_base_parser():
         '--version',
         action='version',
         version=__version__,
-        help='Show version',
+        help='Show version.',
     )
     parser.add_argument(
         '--loglevel',
         type=str,
         choices=['DEBUG', 'INFO', 'CRITICAL', 'NOTSET'],
-        default='INFO',
+        default=os.environ.get('JCLOUD_LOGLEVEL', 'INFO'),
         help='Set the loglevel of the logger',
     )
     return parser
@@ -33,7 +35,7 @@ def set_simple_parser(parser=None):
     parser.add_argument(
         'flow',
         type=str,
-        help='The string ID of a flow',
+        help='The string ID of a flow.',
     )
     return parser