+++ tests/integration/envvars/test_envvars_context_syntax.py
@@ -1,20 +1,21 @@
 import os
+from typing import Dict
 
-import pytest
 from jcloud.flow import CloudFlow
 from jina import Client, DocumentArray
 
-cur_dir = os.path.dirname(os.path.abspath(__file__))
+projects_dir = os.path.join(
+    os.path.dirname(os.path.abspath(__file__)), '..', 'projects'
+)
 
 
-def sorted_dict(d):
+def sorted_dict(d: Dict):
     return dict(sorted(d.items()))
 
 
-@pytest.mark.skip('non-interactive login not supported yet')
 def test_envvars_context_syntax():
     with CloudFlow(
-        path=os.path.join(cur_dir, 'projects', 'envvars_context_syntax'),
+        path=os.path.join(projects_dir, 'envvars_context_syntax'),
         name='context-syntax',
     ) as flow:
         da = Client(host=flow.gateway).post(on='/', inputs=DocumentArray.empty(2))