+++ tests/integration/envvars/test_envvars_default_file.py
@@ -4,17 +4,18 @@ import pytest
 from jcloud.flow import CloudFlow
 from jina import Client, DocumentArray
 
-cur_dir = os.path.dirname(os.path.abspath(__file__))
+projects_dir = os.path.join(
+    os.path.dirname(os.path.abspath(__file__)), '..', 'projects'
+)
 
 
 def sorted_dict(d):
     return dict(sorted(d.items()))
 
 
-@pytest.mark.skip('non-interactive login not supported yet')
 def test_envvars_default_file():
     with CloudFlow(
-        path=os.path.join(cur_dir, 'projects', 'envvars_default_file'),
+        path=os.path.join(projects_dir, 'envvars_default_file'),
         name='default-env',
     ) as flow:
         da = Client(host=flow.gateway).post(on='/', inputs=DocumentArray.empty(2))