+++ tests/integration/test_multi_executors.py
@@ -1,17 +1,15 @@
 import os
 
-import pytest
 from jcloud.flow import CloudFlow
 from jina import Client, DocumentArray
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 
 
-@pytest.mark.skip('non-interactive login not supported yet')
-def test_project2():
+def test_project_multi_executors():
     with CloudFlow(
-        path=os.path.join(cur_dir, 'projects', 'testproject2'),
-        name='testproject2',
+        path=os.path.join(cur_dir, 'projects', 'multi_executors'),
+        name='multi_executors',
     ) as flow:
         da: DocumentArray = Client(host=flow.gateway).post(
             on='/',