+++ tests/integration/test_simple.py
@@ -1,17 +1,15 @@
 import os
 
-import pytest
 from jcloud.flow import CloudFlow
 from jina import Client, DocumentArray
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 
 
-@pytest.mark.skip('non-interactive login not supported yet')
-def test_project1():
+def test_project_simple():
     with CloudFlow(
-        path=os.path.join(cur_dir, 'projects', 'testproject1'),
-        name='testproject1',
+        path=os.path.join(cur_dir, 'projects', 'simple'),
+        name='simple',
     ) as flow:
         da = Client(host=flow.gateway).post(
             on='/',